Snippet Manager Server Implementing user authentication ..4

+++ server/routers/snippetRouter.ts
@@ -1,26 +1,27 @@
-import express, {response, Router} from 'express';
+import express, {Request, response, Router} from 'express';
 import  {Snippet} from '../models/snippetModel';
 import {Document} from "mongoose";
+import {auth} from "../middleware/auth";
 
 const router: Router = express.Router();
 
-router.get('/', async (request, response) => {
+router.get('/', auth, async (request, response) => {
     try {
-        const snippets = await Snippet.find();
+        const snippets = await Snippet.find({ user: request.user });
         response.json(snippets);
     } catch (error) {
         response.status(500).send();
     }
 });
 
-router.post('/', async (request, response) => {
+router.post('/', auth, async (request, response) => {
     try {
         const { title, description, code } = request.body;              // deconstructing a body json object
         if (!description && !code) {
             return response.status(400).json({errorMessage: 'You need to enter at least a description or some code.'});
         }
         const newSnippet = new Snippet({
-            title, description, code
+            title, description, code, user: request.user
         });
         const savedSnippet: Document = await newSnippet.save();
         response.json(savedSnippet);
@@ -30,7 +31,7 @@ router.post('/', async (request, response) => {
 
 });
 
-router.put('/:id', async (request, response) => {
+router.put('/:id', auth, async (request, response) => {
     try {
         const {title, description, code} = request.body;
         const snippetId = request.params.id;       // 608d76defd62c83f63313007
@@ -44,6 +45,9 @@ router.put('/:id', async (request, response) => {
         if (!originalSnippet) {
             return response.status(400).json({errorMessage: `There is no snippet with id ${snippetId}`});
         }
+        if (originalSnippet.user.toString() !== request.user) {
+            return response.status(401).json({ errorMessage: 'Unauthorized' });
+        }
         originalSnippet.title = title;
         originalSnippet.description = description;
         originalSnippet.code = code;
@@ -55,7 +59,7 @@ router.put('/:id', async (request, response) => {
 });
 
 // DELETE http://localhost:5000/snippet/608d76defd62c83f63313007
-router.delete('/:id', async (request, response) => {
+router.delete('/:id', auth, async (request, response) => {
     try {
         const snippetId = request.params.id;       // 608d76defd62c83f63313007
         if (!snippetId) {
@@ -64,6 +68,9 @@ router.delete('/:id', async (request, response) => {
         const existingSnippet = await Snippet.findById(snippetId);
         if (!existingSnippet) {
             return response.status(400).json({errorMessage: `There is no snippet with id ${snippetId}`});
+        }
+        if (existingSnippet.user.toString() !== request.user) {
+            return response.status(401).json({ errorMessage: 'Unauthorized' });
         }
         await existingSnippet.deleteOne();
         response.json(existingSnippet);
